Started some changes to routes - for the immediate map upload functionality I'm bypassing the upload endpoint and having index.html make a POST request to /mapView. I'll save /upload for when users are specifically attempting to upload geojsons to their account.
I'm now able to parse the geoJSON that is uploaded to /mapView, although I still have to save it to the server first and then open it again. I'd like to bypass this eventually, maybe by parsing the raw json data on the front end and sending that instead of a file?

Also have been tinkering with the frontend mapView.html, still desperately trying to get geoJSON feature collections to show up on the leaflet map but still hitting a brick wall.